# Conflicts:
#	my_items.php

+++ my_items.php
@@ -74,8 +74,9 @@
 									<span class="category_label"><?=ucfirst($item->category_name)?></span>
 								</div>
 								<img class="item_img" src="<?=$item->imgPath?>" alt="Image of <?=$item->name?>">
+								
+								<button type="submit" class="delete_item" id="<?=$item->itemID?>" name="delete_item" onclick="return confirm('Are you sure you want to delete this item?');">Delete</button>
 
-								<button type="submit" class="delete_item" id="<?=$item->itemID?>" name="delete_item">Delete</button>
 							</div>
 						<?php endforeach; ?>
 						<div id="items_message"></div>
